Moved programming exercise related tests into subfolder
